# [94X] Make Puppi use existing weights in PFCandidates to speed up