Tip from @ahinzmann : turns out the packedPfCandidates already have puppiWeights stored, so no need to recalculate them.
Puppi module goes from ~0.4s/event to ~0.02s/event ! 🚀 

(This also sped up `hotvrPuppi`, ~1.17s/evt -> ~0.6s/evt ... not sure why but yay)